cleanStar: fix identification of bad stars
For a long time there's been a strong difference between the initial
astrometry (before PSF determination) and the revised astrometry (after
PSF determination).  The initial astrometry uses distorted positions
calculated from initial.centroid.sdss and an initial linear Wcs, while
the revised astrometry uses positions from centroid.sdss and the non-linear
Wcs.  Note that the centroids (particularly the masking) can change
between initial.centroid.sdss and centroid.sdss (due to the different
PSF).  It's therefore not appropriate to make initial.flags.pixel.bad the
check for the quality of the centroid: it's applying the flags from the
initial centroids to the revised centroids, but the flagging is not the
same.  We need to use the flag appropriate to each centroid.

Also removed a bunch of meaningless code that had been copied elsewhere.