Add unit test for fitTAN and improve unit test for fitTANSIP
The unit test for fitTANSIP was expecting the SIP terms to take out
far too much offset and scale change. Now the test only fails
for a quadratic distortion. I'm still puzzled by that failure.

The new test of fitTAN fails even when no distortion is added.
I'm not sure if the problem is in the test or the code being tested.
(And yes, there is far too much duplication between these tests).